refactor: moving llm judge inside server dir
